Remove some forgotten commented code; add a comment of explanation.
We used to set the titleLabel property on the UIButton/FMStepperButton
to show +/-. There were vestiges of that.

It was just too annoying to deal with the baseline of the text in the
UILabel, since +/- symbols in all iOS fonts sit on the baseline.

So now we draw the +/- symbols with UIBezierPaths.